migrate files command
